+++ mkdocs/plugins.py
@@ -3,22 +3,16 @@
 from __future__ import annotations
 
 import logging
-import sys
+from collections.abc import Callable, MutableMapping
+from importlib.metadata import EntryPoint, entry_points
 from typing import (
     TYPE_CHECKING,
     Any,
-    Callable,
     Generic,
     Literal,
-    MutableMapping,
     TypeVar,
     overload,
 )
-
-if sys.version_info >= (3, 10):
-    from importlib.metadata import EntryPoint, entry_points
-else:
-    from importlib_metadata import EntryPoint, entry_points
 
 if TYPE_CHECKING:
     import jinja2.environment
@@ -41,7 +35,9 @@ if TYPE_CHECKING:
     from mkdocs.utils.templates import TemplateContext
 
 if TYPE_CHECKING:
-    from typing_extensions import Concatenate, ParamSpec
+    from typing import Concatenate
+
+    from typing_extensions import ParamSpec
 else:
     ParamSpec = TypeVar