+++ mkdocs/utils/rendering.py
@@ -1,12 +1,13 @@
 from __future__ import annotations
 
 import copy
-from typing import TYPE_CHECKING, Callable
+from typing import TYPE_CHECKING
 
 import markdown
 import markdown.treeprocessors
 
 if TYPE_CHECKING:
+    from collections.abc import Callable
     from xml.etree import ElementTree as etree
 
 # TODO: This will become unnecessary after min-versions have Markdown >=3.4